vendorsetup.sh:  Patch tree on unknown $1 when non-interactive
When invoked interactively it is desirable for
a |. build/envsetup.sh unknown_arg| to emit an error message.

However when invoked non-interactively vendorsetup.sh will
inherit the positional parameters of the caller.  In such a
case when |. build/envsetup.sh| is invoked it is likely that
$1 will be unrecognized.  In an attempt to be the most
compatible with existing scripts that invoke build/envsetup.sh,
vendorsetup.sh now behaves as if $1 was empty and patches the tree.

Change-Id: I645875096191c56b3bbd1102cac6967427388036